// examples/capnweb-demo/src/models/base.ts
// Generated by Typegres introspect
import * as Types from "typegres";
import { Table } from "typegres";

export const Todo = Table("Todo", {
  completed: Types.Bool<1>,
  content: Types.Text<1>,
  createdAt: Types.Timestamp<1>,
  dueDate: Types.Timestamp<0 | 1>,
  id: Types.Text<1>,
  priority: Types.Int4<1>,
  tags: Types.Array<0 | 1, Types.Text<0 | 1>>,
  updatedAt: Types.Timestamp<1>,
  userId: Types.Text<1>,
});

export const User = Table("User", {
  createdAt: Types.Timestamp<1>,
  email: Types.Text<1>,
  id: Types.Text<1>,
  name: Types.Text<1>,
  updatedAt: Types.Timestamp<1>,
});